+++ api/src/types/core.rs
@@ -52,7 +52,6 @@ pub struct Experiment {
     pub tags: Vec<String>,
     pub created_at: chrono::DateTime<chrono::Utc>,
     pub updated_at: chrono::DateTime<chrono::Utc>,
-    pub available_metrics: Vec<String>,
     pub workspace_id: Option<String>,
     pub url: String,
 }