+++ python/tora/__init__.py
@@ -22,7 +22,7 @@ from ._wrapper import (
     tresult,
 )
 
-__version__ = "0.0.9"
+__version__ = "0.0.10"
 
 __all__ = [
     "Tora",